- More work around the shader instance class -- It's now used as a smart pointer in the attributes -- Shorted class name - New ref-counted smart pointers - 64 bit build ! ..and related fixes for it

+++ DSystem/src/DUtils.cpp
@@ -117,7 +117,7 @@ void StrStripBeginEndWhite( char *pStr )
 	if NOT( len )
 		return;
 
-	int	newLen = len;
+	int	newLen = (int)len;
 	for (int i=(int)len-1; i >= 0; --i)
 	{
 		char ch = pStr[i];
